requirements.txt
asgiref==3.5.0
Django==4.0.4
pip==22.0.4
psycopg2==2.9.3
setuptools==61.2.0
sqlparse==0.4.2
tzdata==2022.1
wheel==0.37.1
Pillow==9.1.0
django-durationwidget==1.0.5
djangorestframework==3.13.1